refactor: address PR #358 review — split history, fix docstring, extract helpers
- Extract submission/history functions from _helpers.py into new history.py (269 lines)
- _helpers.py trimmed from 375 to 165 lines
- Fix orchestrator docstring: Three-phase → Four-phase
- Extract _collect_enrichment_coords and _build_order_lookups from orchestrator
- orchestrator.py reduced from 322 to 300 lines
- All 9 pipeline modules now ≤300 lines per #357 constraint
- Update imports in diagnostics, submission, and test files
- 775 tests pass, lint clean, pyright clean
